requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
autopep8==2.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
Django==4.1.3
django-constance==2.9.1
django-picklefield==3.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
httplib2==0.21.0
idna==3.4
oauth2client==4.1.3
packaging==21.3
Pillow==9.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyparsing==3.0.9
pytz==2022.6
redis==4.3.5
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12
